+++ TestProject/Test/WebRequest/Parts/PartUnitTest.cs
@@ -39,7 +39,7 @@ public class PartUnitTest : IClassFixture<WebApplicationFactory<Program>>
         HttpClient client = _factory.CreateClient();
         PartDataLayer dataLayer = new(client);
 
-        int count = await dataLayer.CountAsync();
+        long count = await dataLayer.CountAsync();
         Assert.True(count > 0);
     }
